+++ src/ui/modals/keybinds.rs
@@ -123,16 +123,25 @@ impl Modal for KeybindsModal {
         let keybinds = &app.config.keybinds;
         let header_style = app.config.theme.current_item_style;
 
-        let global = keybinds.global.to_str().collect_vec();
-        let navigation = keybinds.navigation.to_str().collect_vec();
-        let albums = keybinds.albums.to_str().collect_vec();
-        let artists = keybinds.artists.to_str().collect_vec();
-        let directories = keybinds.directories.to_str().collect_vec();
-        let playlists = keybinds.playlists.to_str().collect_vec();
-        let search = keybinds.search.to_str().collect_vec();
-        let queue = keybinds.queue.to_str().collect_vec();
+        let mut global: Vec<_> = keybinds.global.to_str().collect();
+        let mut navigation: Vec<_> = keybinds.navigation.to_str().collect();
+        let mut albums: Vec<_> = keybinds.albums.to_str().collect();
+        let mut artists: Vec<_> = keybinds.artists.to_str().collect();
+        let mut directories: Vec<_> = keybinds.directories.to_str().collect();
+        let mut playlists: Vec<_> = keybinds.playlists.to_str().collect();
+        let mut search: Vec<_> = keybinds.search.to_str().collect();
+        let mut queue: Vec<_> = keybinds.queue.to_str().collect();
 
-        let rows = row_header(&navigation, "Global", header_style)
+        global.sort_by_key(|(key, _, _)| key.to_lowercase());
+        navigation.sort_by_key(|(key, _, _)| key.to_lowercase());
+        albums.sort_by_key(|(key, _, _)| key.to_lowercase());
+        artists.sort_by_key(|(key, _, _)| key.to_lowercase());
+        directories.sort_by_key(|(key, _, _)| key.to_lowercase());
+        playlists.sort_by_key(|(key, _, _)| key.to_lowercase());
+        search.sort_by_key(|(key, _, _)| key.to_lowercase());
+        queue.sort_by_key(|(key, _, _)| key.to_lowercase());
+
+        let rows = row_header(&global, "Global", header_style)
             .into_iter()
             .chain(row(&global, key_area.width, action_area.width, desc_area.width))
             .chain(row_header(&navigation, "Navigation", header_style))
@@ -145,10 +154,10 @@ impl Modal for KeybindsModal {
             .chain(row(&directories, key_area.width, action_area.width, desc_area.width))
             .chain(row_header(&playlists, "Playlists", header_style))
             .chain(row(&playlists, key_area.width, action_area.width, desc_area.width))
-            .chain(row_header(&albums, "Albums", header_style))
-            .chain(row(&queue, key_area.width, action_area.width, desc_area.width))
             .chain(row_header(&search, "Search", header_style))
             .chain(row(&search, key_area.width, action_area.width, desc_area.width))
+            .chain(row_header(&queue, "Queue", header_style))
+            .chain(row(&queue, key_area.width, action_area.width, desc_area.width))
             .collect_vec();
 
         self.scrolling_state.set_content_len(Some(rows.len()));